requirements.txt
Babel==2.5.1
beautifulsoup4==4.6.0
bleach==2.1.1
Cartridge==0.12.0
certifi==2017.7.27.1
chardet==3.0.4
codecov==2.0.15
coverage==4.5.1
dj-database-url==0.5.0
Django==1.10.8
django-admin-rangefilter==0.3.5
django-advanced-filters==1.0.7.1
django-appconf==1.0.2
django-braces==1.4.0
django-cities==0.5.0.6
django-compressor==2.2
django-contrib-comments==1.8.0
django-debug-toolbar==1.8
django-environ==0.4.4
django-excel==0.0.10
django-extensions==1.9.1
django-extra-views==0.6.4
django-filter==1.0.4
django-guardian==1.4.1
django-haystack==2.6.1
django-model-utils==3.1.1
django-mptt==0.8.7
django-mptt-admin==0.5.1
django-oscar==1.5
django-phonenumber-field==1.3.0
django-tables2==1.11.0
django-treebeard==4.1.2
django-userena==2.0.1
django-widget-tweaks==1.4.1
django-zurb-foundation==5.5.0
dnspython==1.15.0
easy-thumbnails==2.4.2
email-validator==1.0.2
et-xmlfile==1.0.1
factory-boy==2.9.2
Faker==0.8.3
filebrowser-safe==0.4.7
flake8==3.5.0
future==0.16.0
grappelli-safe==0.4.7
gunicorn==19.8.1
html2text==2014.12.29
html5lib==0.999999999
idna==2.6
jdcal==1.4
Jinja2==2.10
lml==0.0.1
lxml==4.2.1
MarkupSafe==1.0
mccabe==0.6.1
Mezzanine==4.2.3
mock==2.0.0
mysqlclient==1.3.12
nose==1.3.7
numpy==1.14.0
oauthlib==2.0.4
olefile==0.44
openpyxl==2.5.3
pandas==0.22.0
pbr==3.1.1
phonenumberslite==8.8.1
Pillow==4.3.0
purl==1.3.1
pycodestyle==2.3.1
pycountry==17.5.14
pyexcel==0.5.8
pyexcel-ezodf==0.3.4
pyexcel-handsontable==0.0.1
pyexcel-io==0.5.6
pyexcel-ods3==0.5.2
pyexcel-pygal==0.0.2
pyexcel-webio==0.1.4
pyexcel-xls==0.5.6
pyexcel-xlsx==0.5.6
pyflakes==1.6.0
pygal==2.4.0
PyPDF2==1.26.0
python-dateutil==2.6.1
pytz==2017.2
rcssmin==1.0.6
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
rjsmin==1.0.12
selenium==3.9.0
simplejson==3.6.5
six==1.11.0
sorl-thumbnail==12.4a1
sqlparse==0.2.3
swapper==1.1.0
texttable==1.2.1
tqdm==4.19.5
tzlocal==1.4
Unidecode==0.4.21
urllib3==1.22
webencodings==0.5.1
whitenoise==3.3.0
Whoosh==2.7.4
xhtml2pdf==0.0.6
xlrd==1.1.0
xlwt==1.3.0
psycopg2==2.7.2